+ Spring source format applied + BaseEntity version property changed from Long to Integer + Extracted enumClass constant for AnnotatedEnumUserType

+++ arena-hibernate/trunk/arena-hibernate/src/main/java/org/openinsula/arena/hibernate/entity/BaseEntity.java
@@ -20,44 +20,44 @@ public abstract class BaseEntity<T extends Number> implements Serializable {
 	@Id
 	@GeneratedValue
 	private T id;
-	
+
 	@Version
-	private Long version;
-	
+	private Integer version;
+
 	public T getId() {
 		return id;
 	}
-	
+
 	public void setId(final T id) {
 		this.id = id;
 	}
-	
-	public void setVersion(final Long version) {
+
+	public void setVersion(final Integer version) {
 		this.version = version;
 	}
-	
-	public Long getVersion() {
+
+	public Integer getVersion() {
 		return version;
 	}
-	
+
 	@Override
 	public String toString() {
 		return String.format("[%s: id= %s]", getClass().getName(), ObjectUtils.nullSafeToString(id));
 	}
-	
+
 	@Override
 	public boolean equals(final Object obj) {
 		if (this == obj) {
 			return true;
 		}
-		
+
 		if (obj == null || !getClass().isAssignableFrom(obj.getClass())) {
 			return false;
 		}
-		
+
 		return ObjectUtils.nullSafeEquals(id, ((BaseEntity<?>) obj).id);
 	}
-	
+
 	@Override
 	public int hashCode() {
 		return ObjectUtils.nullSafeHashCode(id);
